Move functionality from main.rs to respective modules.
