+++ src/aes_encryption.rs
@@ -1,7 +1,9 @@
 use aes_gcm::{aead::{Aead, AeadCore, KeyInit, OsRng}, Aes256Gcm, Nonce, Key};
+use crate::key_generation;
+use crate::file_management::{read_file, clear_write_file};
 
 
-pub fn encrypt(data: &[u8], key: &[u8; 32]) -> Result<EncryptedData, aes_gcm::Error> {
+fn encrypt(data: &[u8], key: &[u8; 32]) -> Result<EncryptedData, aes_gcm::Error> {
     // Transformed from a byte array:
     let key: &Key<Aes256Gcm> = key.into();
 
@@ -13,7 +15,7 @@ pub fn encrypt(data: &[u8], key: &[u8; 32]) -> Result<EncryptedData, aes_gcm::Er
     Ok(EncryptedData::new(&ciphertext, &nonce))
 }
 
-pub fn decrypt(data: EncryptedData, key: &[u8; 32]) -> Result<Vec<u8>, aes_gcm::Error> {
+fn decrypt(data: EncryptedData, key: &[u8; 32]) -> Result<Vec<u8>, aes_gcm::Error> {
     let key: &Key<Aes256Gcm> = key.into();
     let cipher = Aes256Gcm::new(key);
     
@@ -24,7 +26,54 @@ pub fn decrypt(data: EncryptedData, key: &[u8; 32]) -> Result<Vec<u8>, aes_gcm::
 }
 
 
-pub struct EncryptedData {
+pub fn encrypt_file(path: &str, passphrase: String) {
+    let (bytes_read, file_content) = read_file(path)
+        .expect("Failed to read file.");
+
+    println!("Bytes to encrypt: {}", bytes_read);
+
+    let file_content = file_content.as_slice();
+    let salt = key_generation::generate_salt();
+    let key = key_generation::derive_key_from_passphrase(passphrase.as_bytes(), &salt);
+
+    let encrypted_data = encrypt(file_content, &key).expect("An error occurred when encrypting data.");
+
+    // Add the Nonce at the beginning of the encrypted data for later retrieval.
+    let mut new_file_content = encrypted_data.nonce().clone();
+    new_file_content.append(&mut salt.to_vec()); // Add the used salt
+    new_file_content.append(encrypted_data.data().clone().as_mut());
+
+    clear_write_file(path, new_file_content).expect("Failed to write encrypted data to file.");
+}
+
+pub fn decrypt_file(path: &str, passphrase: String) {
+    let (bytes_read, file_contents) = read_file(path).unwrap();
+    println!("Bytes read: {}", bytes_read);
+    let mut file_contents = file_contents;
+
+    // Get the nonce from the first 12 bytes of the file.
+    let nonce_drain = file_contents.drain(0..12);
+    let mut nonce: Vec<u8> = Vec::new();
+    for byte in nonce_drain {
+        nonce.push(byte);
+    }
+
+    // Get the salt from the now first 16 bytes of the file.
+    let salt_drain = file_contents.drain(0..16);
+    let mut salt: Vec<u8> = Vec::new();
+    for byte in salt_drain {
+        salt.push(byte);
+    }
+
+    let encrypted_data = EncryptedData::new(file_contents.as_slice(), nonce.as_slice());
+    let key = key_generation::derive_key_from_passphrase(passphrase.as_bytes(), salt.as_slice());
+    let decrypted_data = decrypt(encrypted_data, &key).unwrap();
+
+    clear_write_file(path, decrypted_data).expect("Failed to write encrypted contents to file.");
+}
+
+
+struct EncryptedData {
     data: Vec<u8>,
     nonce: Vec<u8>,
 }
